Create a "request" type for requests rather than using c_pointer.
It is more typesafe to use a Mercury data type for each foreign type, it
prevents any user from mixing up different c pointer.  I've also returned
the request inside a maybe() when it is created by fcgx_init_request which
is more idomatic and also helps with safety.

This change has also allowed me to remove the type casts in the C code as
that's handled automatically by Mercury's FFI.

multi.m:
    As above.